feat: refactor middleware and models for settings, user, and admin; create project model
